Move configuration.js under node-manager folder. Update references.
